Added 3D test for labels
